予定の色を変更
