+++ src/syntax/html-syntax-generator.ts
@@ -1,7 +1,7 @@
 import { Member, MyGroupEvent, ScheduleEvent } from '../garoon/schedule';
 import { DateTime, formatDateTime } from '../utils/date-time';
 import { AbstractSyntaxGenerator } from './base/abstract-syntax-generator';
-import { COLOR, getEventMenuColorCode } from './colors';
+import { getEventMenuColorCode } from './event-menu-color';
 
 export class HtmlSyntaxGenerator extends AbstractSyntaxGenerator {
     createTitle(dateTime: DateTime) {
@@ -54,7 +54,7 @@ export class HtmlSyntaxGenerator extends AbstractSyntaxGenerator {
     }
 
     private createMembers(members: Member[]) {
-        return `<span style="color: ${COLOR.BROWN};">(${members
+        return `<span style="color: #b99976;">(${members
             .map((member) => member.name)
             .join(', ')
             .replace(/, $/, '')})</span>`;
